+++ src/Service/Magazine/MagazineStructure.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Service\Magazine;
+
+final readonly class MagazineStructure
+{
+    /**
+     * @param array<int, array<mixed>> $categoryTags
+     * @param string[] $chapterCoordinates
+     */
+    public function __construct(
+        public array $categoryTags,
+        public array $chapterCoordinates,
+        public ?string $frontPageArticleCoordinate,
+    ) {
+    }
+}